insightface_service/camera_stream/requirements.txt
ffmpeg-python
ultralytics
opencv-python-headless
numpy
requests
PyYAML
pillow
insightface
onnxruntime
# db clients según tu BD, por ejemplo:
# psycopg2-binary
# mysql-connector-python